Add tsung as dep.
Start tsung_controller node and application.

Closes #11
